Several improvements in the help page

+++ src/odbase/var/www/cgi-bin/help.sh
@@ -1,19 +1,27 @@
 #!/bin/sh
 echo "Content-type: text/html"
 echo
+lang=`cat /etc/opendomo/lang`
 TOPIC=`echo $QUERY_STRING | cut -f1 -d.`
 echo "<html><head><link rel='stylesheet' type='text/css' media='screen' href='/cgi-bin/css.cgi?admin'></head><body>"
-if test -f /usr/local/opendomo/docs/$TOPIC.txt; then
-	echo "<fieldset>"
-	cat /usr/local/opendomo/docs/$TOPIC | sed 's/^/# /'
-	echo "</fieldset>"
-else
-	echo "<fieldset>"
-	echo "<legend>$TOPIC</legend>"
-	i18n.sh "Manual page not found"
-	echo "</fieldset>"
+echo "<fieldset>"
+if test -f /usr/local/opendomo/docs/$TOPIC.$lang.txt; then
+	cat /usr/local/opendomo/docs/$TOPIC.$lang.txt 
+else 
+	if test -f /usr/local/opendomo/docs/$TOPIC.txt; then
+		cat /usr/local/opendomo/docs/$TOPIC.txt 
+	else
+		echo "<legend>$TOPIC</legend>"
+		i18n.sh "Manual page not found"
+	fi
 fi
-echo "<div class='toolbar'><a href='http://www.opendomo.com/wiki/index.php?title=$TOPIC' target='_new'>"
+echo "</fieldset>"
+echo "<div class='toolbar'>"
+echo "<a class='button' href='http://www.opendomo.com/wiki/index.php?title=$TOPIC' target='_new'>"
 i18n.sh "More information"
-echo "</a></div>"
+echo "</a>"
+echo "<a class='button' href='javascript:hideBubble()'>"
+i18n.sh "Close"
+echo "</a>"
+echo "</div>"
 echo "</body></html>"
